RING-28616 GET: extract CorruptSafeStream into its own file
